requirements.txt
uvicorn==0.29.0
python-dotenv==1.0.1
fastapi==0.110.1
ruff==0.3.5
httpx==0.27.0
numpy==1.26.4
pytest==8.1.1
matplotlib==3.8.4
pytest-cov==5.0.0
pytest-asyncio==0.23.6